Rename from test_config.py to config.py
