+++ localstack-core/localstack/testing/config.py
