+++ tests/aws/services/s3control/test_s3control.py
@@ -5,8 +5,8 @@ from botocore.client import Config
 from botocore.exceptions import ClientError
 
 from localstack.testing.aws.util import is_aws_cloud
+from localstack.testing.config import TEST_AWS_ACCESS_KEY_ID, TEST_AWS_SECRET_ACCESS_KEY
 from localstack.testing.pytest import markers
-from localstack.testing.test_config import TEST_AWS_ACCESS_KEY_ID, TEST_AWS_SECRET_ACCESS_KEY
 from localstack.utils.strings import short_uid
 from localstack.utils.urls import localstack_host
