+++ tests/aws/services/sts/test_sts.py
@@ -7,8 +7,8 @@ import requests
 from localstack import config
 from localstack.constants import APPLICATION_JSON
 from localstack.testing.aws.util import create_client_with_keys
+from localstack.testing.config import TEST_AWS_ACCESS_KEY_ID
 from localstack.testing.pytest import markers
-from localstack.testing.test_config import TEST_AWS_ACCESS_KEY_ID
 from localstack.utils.aws.request_context import mock_aws_request_headers
 from localstack.utils.numbers import is_number
 from localstack.utils.strings import short_uid, to_str
