+++ tests/aws/test_serverless.py
@@ -5,8 +5,8 @@ import os
 import pytest
 
 from localstack.testing.aws.util import is_aws_cloud
+from localstack.testing.config import TEST_AWS_ACCESS_KEY_ID
 from localstack.testing.pytest import markers
-from localstack.testing.test_config import TEST_AWS_ACCESS_KEY_ID
 from localstack.utils.aws import arns
 from localstack.utils.common import retry, run
 from localstack.utils.testutil import get_lambda_log_events
