+++ tests/unit/aws/test_connect.py
@@ -20,7 +20,7 @@ from localstack.constants import INTERNAL_AWS_SECRET_ACCESS_KEY
 from localstack.http import Response
 from localstack.http.duplex_socket import enable_duplex_socket
 from localstack.http.hypercorn import GatewayServer
-from localstack.testing.test_config import TEST_AWS_ACCESS_KEY_ID
+from localstack.testing.config import TEST_AWS_ACCESS_KEY_ID
 from localstack.utils.aws.client_types import ServicePrincipal
 from localstack.utils.aws.request_context import extract_access_key_id_from_auth_header
 from localstack.utils.net import get_free_tcp_port
