+++ tests/unit/test_apigateway.py
@@ -40,7 +40,7 @@ from localstack.services.apigateway.templates import (
     ResponseTemplates,
     VelocityUtilApiGateway,
 )
-from localstack.testing.test_config import TEST_AWS_REGION_NAME
+from localstack.testing.config import TEST_AWS_REGION_NAME
 from localstack.utils.aws.aws_responses import requests_response
 from localstack.utils.common import clone
