+++ tests/unit/test_dynamodb.py
@@ -9,7 +9,7 @@ from localstack.services.dynamodb.utils import (
     SchemaExtractor,
     dynamize_value,
 )
-from localstack.testing.test_config import (
+from localstack.testing.config import (
     TEST_AWS_ACCESS_KEY_ID,
     TEST_AWS_ACCOUNT_ID,
     TEST_AWS_REGION_NAME,
